// packages/glyphengine/src/interfaces/iGlyph.ts
export interface IGlyph {
  label: number;
  parent: number;
  positionX: number;
  positionY: number;
  positionZ: number;
  rotationX: number;
  rotationY: number;
  rotationZ: number;
  scaleX: number;
  scaleY: number;
  scaleZ: number;
  colorR: number;
  colorG: number;
  colorB: number;
  alpha: number;
  geometry: number;
  topology: number;
  ratio: number;
  rotationRateX: number;
  rotationRateY: number;
  rotationRateZ: number;
  tag: string;
  url: string;
  desc: string;
}